% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chencov.cfr.R
\name{chencov.cfr}
\alias{chencov.cfr}
\title{Obtain CFR In Individual Survival Data}
\usage{
chencov.cfr(object, covv)
}
\arguments{
\item{object}{a model object from chencov.p.}

\item{covv}{covariate value.}
}
\value{
A list containing the following components:
\item{method}{the source.}
\item{estimation}{the estimate at the study time,which contains Time(study time),CFR and Var(variance)}
}
\description{
Obtain CFR with covariate value for chencov.p. See references for details.
}
\examples{
data(individual)
#get individual data
sudata<-surdata(data=individual,t=60,type=1)
#data for calculation,study time:60
## Not run:
obj<-chencov.p(sudata,cov=c("sex","age"),p=c(0.2,0.4,0.5,0.4,0.6))
## End(Not run)
#cov:covariate; p:starting parameter values.
## Not run:
chencov.cfr(object=obj,covv=c(0,1))
## End(Not run)
#calculating CFR with covariate value c(0,1);sex=0 for male,age=1 for more than 60 years old.
}
\references{
Zheng Chen. Estimating the case fatality raye using a constant cure-death hazard ration. Lifetime Data Anal. 2009,15,316-329.
}
