% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chencov.p.R
\name{chencov.p}
\alias{chencov.p}
\title{Estimate Coefficient values In Individual Survival Data}
\usage{
chencov.p(data, cov, p)
}
\arguments{
\item{p}{starting parameter values;p[1] for beta0,p[2:(n+1)] for beta1 with respect to death,p[(n+2):(2*n+1)] for beta2 with respect to cure.}

\item{cov}{covariate}

\item{data}{individual survival data transformed by surdata.}
}
\value{
A list containing the following components:
\item{method}{the source.}
\item{cov}{ covariate.}
\item{estimate}{ coefficient values.}
\item{var}{ variance matrix}
\item{time}{ study time}
\item{deaths}{ number of death}
}
\description{
Estimate coefficient values of covariate.
}
\examples{
data(individual)
#get individual data
sudata<-surdata(data=individual,t=60,type=1)
#data for calculation
## Not run:
chencov.p(sudata,cov=c("sex","age"),p=c(0.2,0.4,0.5,0.4,0.6))
## End(Not run)
#cov:covariate; p:starting parameter values.
}
\references{
Zheng Chen. Estimating the case fatality raye using a constant cure-death hazard ration. Lifetime Data Anal. 2009,15,316-329.
}
