% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jewell.cfr.R
\name{jewell.cfr}
\alias{jewell.cfr}
\title{Modified KM Method For CFR In Individual Survival Data}
\usage{
jewell.cfr(data)
}
\arguments{
\item{data}{individual survival data transformed by surdata.}
}
\value{
A list containing the following components:
\describe{
\item{method}{ the source.}
\item{CFR_a}{the interpretation of method A.}
\item{CFR_b}{the interpretation of method B.}
\item{estimation}{the estimate at the study time,which contains method,Time(study time),CFR,Greedvar(Greedwood variance) and CI.}
}
}
\description{
Calculate CFR using modified Kaplan-Meier(KM) method, including method-A(death cumulative incidence) and method-B(death cumulative incidence divided by death and cure cumulative incidence). See references for details
}
\examples{
data(individual)
data1<-surdata(data=individual,t=60,type=1)
jewell.cfr(data1)
#calculate CFR
}
\references{
Nicholas P.Jewell. Non-parametric estimation of the case fatality ratio with competing risks data:An application to Severe Acute Respiratory Syndrome(SARS), Statist.Med. 2007,26,1982-1998.
}
