% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.donga.R
\name{out.donga}
\alias{out.donga}
\title{Calculating CFR Using Parametric Method In Individual Data}
\usage{
out.donga(data, p)
}
\arguments{
\item{data}{individual survival data transformed by surdata.}

\item{p}{starting parameter values. p[1:2] for the mean and variance of death,p[3:4] for the mean and variance of cure,p[5] for CFR.}
}
\value{
A list containing the following components:
\item{method}{the source.}
\item{estimation}{the estimate at the study time,which contains Time(study time),CFR,Var(variance) and CI.}
\item{p}{parameter values}
\item{Var}{variance matrix}
}
\description{
Calculating CFR using gamma parametric method in individual survival data. See references for details.
}
\examples{
data(individual)
#get data
data1<-surdata(data=individual,t=60,type=1)
##survival data for calculation
}
\references{
Christl A Donnelly. Epidemiological determinants of spread of causal agent of severe acute respiratory syndrome in Hong Kong. The Lancet,2003,361:1761-1832.
}
