% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.donkm.R
\name{out.donkm}
\alias{out.donkm}
\title{Calculating CFR Using KM method In Individual Data}
\usage{
out.donkm(data, p)
}
\arguments{
\item{data}{individual survival data transformed by surdata.}

\item{p}{starting parameter values. p[1:2] for the mean and variance of death,p[3:4] for the mean and variance of cure,p[5] for CFR.}
}
\value{
A list containing the following components:
\item{method}{the source.}
\item{estimation}{the estimate at the study time,which contains Time or study time,CFR,Var orvariance and CI.}
\item{p}{parameter values}
\item{Var}{variance matrix}
}
\description{
Calculating CFR using KM method in individual survival data. See references for details.
}
\references{
Christl A Donnelly. Epidemiological determinants of spread of causal agent of severe acute respiratory syndrome in Hong Kong. The Lancet,2003,361:1761-1832.
}
