% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdata.wei.R
\name{sdata.wei}
\alias{sdata.wei}
\title{Simulate Individual Data Using Weibull Distribution}
\usage{
sdata.wei(num, cfr, para, seed = Sys.Date())
}
\arguments{
\item{num}{number of size.}

\item{cfr}{the set cfr.}

\item{para}{the parameter values. para[1:2] for shape and scale of death,para[3:4] for shape and scale of cure.}

\item{seed}{the default seed is current time.}
}
\value{
A list containing the following components:
\item{id}{patient id number.}
\item{time1}{entry time.}
\item{time2}{the time of cure or death.}
\item{time}{survival time.}
\item{dc}{indicator of outcome.1-death,2-cure.}
}
\description{
Simulate individual data using Weibull distribution.
}
