% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yip.cfr.R
\name{Yip.cfr}
\alias{Yip.cfr}
\title{Calculating CFR Using Chain Multinomial In Summarized Data}
\usage{
Yip.cfr(data,t,B)
}
\arguments{
\item{data}{summarized survival data}

\item{t}{the study time.}

\item{B}{the window width of a kernel function.}
}
\value{
A list containing the following components:
\item{CFR}{CFR.}
\item{lower}{the lowe of CFR's  CI}
\item{upper}{the upper of CFR's CI.}
\item{var}{ the variance of CFR}
}
\description{
Calculating CFR chain multinomial in summarized data. See references for details.
}
